Remove interactions with docker-machine
Another assumption about the running environment. Can't garuantee the docker-machine to be used is default, don't have a reliable way of determining this without some sort of config/env/flag. For now, keeping things simple.

Removes the automatic machine starting and the open command.

In theory this should also allow this to run on Linux distros, though that is untested.